# fix(markdown.js): use proposed fixes using negative look-ahead