@marijnh 2025-08-27
I've verified that the first one doesn't solve the problem (the given problematic input is still overly slow to match) and actually changes the meaning of the regexp (removing the optional space). I haven't looked deeply into the other changes but they also don't look very convincing.